+++ examples/configparser/example.py
@@ -10,10 +10,12 @@ class ServerSchema(Schema):
     host = IpAddress("host")
     port = Int("port", default=80)
 
+
 # this schema describes the entire config
 class ConfigSchema(Schema):
     # server schema as sub schema
     server = ServerSchema()
+
 
 # read config
 config = configparser.ConfigParser()