+++ CHANGELOG.md
@@ -3,13 +3,16 @@
 ## [Unreleased]
 ### Added
 ### Fixed
-- some attribute validators still have issues with a None default (#16)
-- For string attributes a blank default value should infer blank=True (#15)
 ### Changed
-
 ### Deprecated
 ### Removed
 ### Security
+
+## [1.1.2]
+
+### Fixed
+- some attribute validators still have issues with a None default (#16)
+- For string attributes a blank default value should infer blank=True (#15)
 
 ## [1.1.0]